Refactor RootShell and disable auto restart of service
