+++ aFWall/src/main/java/dev/ukanth/ufirewall/InterfaceTracker.java
@@ -46,7 +46,7 @@ import java.util.Enumeration;
 import java.util.Iterator;
 
 import dev.ukanth.ufirewall.log.Log;
-import dev.ukanth.ufirewall.service.RootShell.RootCommand;
+import dev.ukanth.ufirewall.service.RootShellService.RootCommand;
 import dev.ukanth.ufirewall.util.G;
 
 public final class InterfaceTracker {
