+++ aFWall/src/main/java/dev/ukanth/ufirewall/preferences/ExpPreferenceFragment.java
@@ -21,7 +21,7 @@ import java.io.File;
 
 import dev.ukanth.ufirewall.Api;
 import dev.ukanth.ufirewall.R;
-import dev.ukanth.ufirewall.service.RootShell;
+import dev.ukanth.ufirewall.service.RootShellService;
 import dev.ukanth.ufirewall.util.G;
 
 public class ExpPreferenceFragment extends PreferenceFragment implements
@@ -125,7 +125,7 @@ public class ExpPreferenceFragment extends PreferenceFragment implements
 							File f = new File(s);
 							if (f.exists() && f.isDirectory()) {
 								//make sure it's executable
-								new RootShell.RootCommand()
+								new RootShellService.RootCommand()
 										.setReopenShell(true)
 										.setLogging(true)
 										.run(ctx, "chmod 755 " + f.getAbsolutePath());
@@ -174,10 +174,10 @@ public class ExpPreferenceFragment extends PreferenceFragment implements
 			File f = new File(s);
 			if (f.exists() && f.isDirectory()) {
 				String filePath  = s + "/" + initScript;
-					new RootShell.RootCommand()
-						.setReopenShell(true).setCallback(new RootShell.RootCommand.Callback() {
+					new RootShellService.RootCommand()
+						.setReopenShell(true).setCallback(new RootShellService.RootCommand.Callback() {
 						@Override
-						public void cbFunc(RootShell.RootCommand state) {
+						public void cbFunc(RootShellService.RootCommand state) {
 							if (state.exitCode == 0) {
 								returnFlag[0] = true;
 							}
