+++ aFWall/src/main/res/values/strings.xml
@@ -290,6 +290,7 @@
     <string name="showuid_summary_on">Show app UID\'s in the main view</string>
     <string name="showuid_summary_off">Hide app UID\'s in the main view. Enabling can be useful to track logs.</string>
     <string name="languages">Languages</string>
+    <string name="change_languages">Change Language</string>
 
     <string name="misc_header">Languages/Plugins</string>
     <!-- 1.2.6 -->
